Unfuck tag borders
Tag border growth algorithm didn't perform lookahead or lookback when
deciding the Y-axis growth.
This causes issues when the next line is shorter than the current line.

Fixed this by introducing well-documented logic for deciding Y-growth
direction, also moved it into UIHelpers because why not.

I'm sure there's some way to make that algorithm not need the
right_align flag, but for now this is how it works and its absolutely
fine.

Fixes #11