Do not bail symbols loop for uncorrelated a/c
Previously, for an uncorrelated target,
the symbol loop would bail immediately.

This bailing point was moved until _after_
the adding of the ac_symbol interaction object
and the drawing of target symbols.

As a result, target symbols for moving STBY planes return, fixes #14
Next, this also makes the interaction object for such planes useable
again, fixes #13

Also moves the metres-to-pixels calculation for target size
outside of the loop, sharing the calculation over all targets.