Add clearance arrows to callsign
When cleared for takeoff, these should show an up arrow.
They do this by hooking the DEPA ground status.

When cleared to land, they should show a downwards arrow.
For this they check the flight strip for the TopSky Mark of Freq marks.

The cleared to line up flag is at this time not implemented

Closes #23