Preallocate and vacuum Patatoides
Instead of a map, use a vector preallocated to the required size.
This should reduce allocations quite a bit.

Also tracks a timestamp of when a callsign was last touched,
deleting its entry if this was a long time ago.
This is of course a tradeoff, having a small overhead,
but should be worth it.