Remove Patatoides
Indirectly fixes #50

Storing the patatoides object takes a lot of memory,
and is slightly unnecessary.
We can fetch the previous positions from EuroScope
and just render aircraft on those.

Downside to this approach is that afterglows keep jiggling due to the
randomness.
Thus, we disable randomness on afterglows.
We could do a workaround where we save the random seed for afterglows
to keep some form of randomness,
but this seems unnecessary.