Use connected acf type, rather than filed
Thanks to the suggestion of the man Lars Bergmann,
we're now using the aircraft type as sent by FSD,
rather than the one in the flightplan.
We can still fall back to flightplan should no FSD info be avilable,
as the mechanism is rather fragile.

Getting this data involves schlepping it from the CPlugin,
to the CRadarScreen, but it works!

Fixes #80